refactor: app widgets are now in the appwidget package.
